Fix #118. Use ServiceLoader to load services
FactoryFinder is very, possibly overly, generic. It looks like it was
written to re-use across multiple projects. I was tempted by a larger
refactoring as well as a greater use of generics but I opted to just fix
the immediate issue. Wider refactorings can follow later if there is
interest in that.