Implement alternative JavaBean support
Add support for multiple JavaBean specification implementations.

If the class java.beans.BeanInfo is available, the full JavaBeans
implementation is used (same behaviour as before this patch).

If the full JavaBeans implementation is not available, a built-in,
stand-alone implementation is used that only provides the JavaBeans
functionality that does not depend on any of the java.beans.* classes
(essentially getter/setter support).

Note that the configuration system properties are intended for testing
only. They are NOT part of the public API.